+++ NEWS.md
@@ -3,3 +3,5 @@
 * Added the `read_ods_multsheets()` function
 
 * Added a `NEWS.md` file to track changes to the package.
+
+* Edited `find_col_info()` to return a nice message instead of an error when there are no more invalid columns left to check.